Updated all the int variables to float variables
This change was mentioned in the previous commit but was not done. This change was done in the Unit, MeleeUnit & RangedUnit classes

+++ Assets/Scripts/Unit.cs
@@ -12,12 +12,12 @@ namespace Assets.Scripts
         //Abstract class with no code implemented 
         //Used for the sake of re-use and inheritance        
         protected Vector3 position;
-        protected int health;
-        protected int maxHealth;
-        protected int speed;
-        protected int attack;
-        protected int attackRange;
-        protected int faction;
+        protected float health;
+        protected float maxHealth;
+        protected float speed;
+        protected float attack;
+        protected float attackRange;
+        protected float faction;
         //protected string symbol;
         protected string name;
         protected bool isAttacking;
